Move 'computeBpmnBaseClassName' out of StyleConfigurator
