+++ src/component/mxgraph/config/StyleConfigurator.ts
@@ -380,8 +380,3 @@ export default class StyleConfigurator {
     return value;
   }
 }
-
-// TODO move to StyleUtils + tests as well
-export function computeBpmnBaseClassName(name: string): string {
-  return !name ? '' : 'bpmn-' + name.replace(/([A-Z])/g, g => '-' + g[0].toLowerCase());
-}
